Push loading of `KnownTypes` into the package level
